requirements.txt
langchain
langchain-community
pypdf
sentence-transformers
transformers
faiss-cpu
torch
PyMuPDF
textwrap3
